+++ R/attach.R
@@ -1,3 +1,3 @@
 .onAttach <- function(...) {
-  packageStartupMessage(sprintf("rRofex v.%s\nAn Open Source initiative by Grupo Rofex", packageVersion("rRofex")))
+  packageStartupMessage(sprintf("rRofex v.%s\nAn Open Source initiative by Matba Rofex", packageVersion("rRofex")))
 }