+++ R/rRofex.R
@@ -5,6 +5,6 @@
 #' @importFrom tidyr separate replace_na pivot_wider
 #' @importFrom magrittr %>%
 #' @importFrom glue glue glue_collapse
-#' @importFrom purrr some modify_depth
+#' @importFrom purrr some modify_depth map_df pluck map
 #' @importFrom rlang is_null
 "_PACKAGE"